Migrate imageviewer example camera view to PlatformCamera abstraction
