--- examples/imageviewer/shared/src/commonMain/kotlin/example/imageviewer/view/CameraView.common.kt
@@ -1,12 +1,0 @@
-package example.imageviewer.view
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import example.imageviewer.PlatformStorableImage
-import example.imageviewer.model.PictureData
-
-@Composable
-expect fun CameraView(
-    modifier: Modifier,
-    onCapture: (picture: PictureData.Camera, image: PlatformStorableImage) -> Unit
-)
